+++ engine/src/execution_graph/logic_controllers/PhysicalPlanGenerator.h
@@ -345,6 +345,18 @@ struct tree_processor {
 		return str;
 	}
 
+	std::string transform_physical_plan(std::string json){
+        std::istringstream input(json);
+        boost::property_tree::ptree p_tree;
+        boost::property_tree::read_json(input, p_tree);
+
+        transform_json_tree(p_tree);
+
+        std::ostringstream output;
+        boost::property_tree::write_json(output, p_tree);
+        return output.str();
+	}
+
 	std::tuple<std::shared_ptr<ral::cache::graph>,std::size_t> build_batch_graph(std::string json) {
 		auto query_graph = std::make_shared<ral::cache::graph>();
 		std::size_t max_kernel_id = 0;
@@ -382,6 +394,11 @@ struct tree_processor {
 			if (it != config_options.end()){
 				join_partition_size_thresh = std::stoull(config_options["JOIN_PARTITION_SIZE_THRESHOLD"]);
 			}
+			int concatenating_cache_num_bytes_timeout = 100000; // 100ms
+			it = config_options.find("CONCATENATING_CACHE_NUM_BYTES_TIMEOUT");
+			if (it != config_options.end()){
+				concatenating_cache_num_bytes_timeout = std::stoi(config_options["CONCATENATING_CACHE_NUM_BYTES_TIMEOUT"]);
+			}
 			
 			auto child_kernel_type = child->kernel_unit->get_type_id();
 			auto parent_kernel_type = parent->kernel_unit->get_type_id();
@@ -397,7 +414,7 @@ struct tree_processor {
 					bool right_concat_all = join_type == ral::batch::LEFT_JOIN || join_type == ral::batch::OUTER_JOIN || join_type == ral::batch::CROSS_JOIN;
 					bool concat_all = index == 0 ? left_concat_all : right_concat_all;
 					cache_settings join_cache_machine_config = cache_settings{.type = CacheType::CONCATENATING, .num_partitions = 1, .context = context->clone(),
-						.concat_cache_num_bytes = join_partition_size_thresh, .concat_all = concat_all};
+						.concat_cache_num_bytes = join_partition_size_thresh, .num_bytes_timeout = concatenating_cache_num_bytes_timeout, .concat_all = concat_all};
 					query_graph.addPair(ral::cache::kpair(child->kernel_unit, parent->kernel_unit, port_name, join_cache_machine_config));					
 
 				} else {
@@ -414,9 +431,9 @@ struct tree_processor {
 					bool left_concat_all = join_type == ral::batch::RIGHT_JOIN || join_type == ral::batch::OUTER_JOIN || join_type == ral::batch::CROSS_JOIN;
 					bool right_concat_all = join_type == ral::batch::LEFT_JOIN || join_type == ral::batch::OUTER_JOIN || join_type == ral::batch::CROSS_JOIN;
 					cache_settings left_cache_machine_config = cache_settings{.type = CacheType::CONCATENATING, .num_partitions = 1, .context = context->clone(),
-						.concat_cache_num_bytes = join_partition_size_thresh, .concat_all = left_concat_all};
+						.concat_cache_num_bytes = join_partition_size_thresh, .num_bytes_timeout = concatenating_cache_num_bytes_timeout, .concat_all = left_concat_all};
 					cache_settings right_cache_machine_config = cache_settings{.type = CacheType::CONCATENATING, .num_partitions = 1, .context = context->clone(),
-						.concat_cache_num_bytes = join_partition_size_thresh, .concat_all = right_concat_all};
+						.concat_cache_num_bytes = join_partition_size_thresh, .num_bytes_timeout = concatenating_cache_num_bytes_timeout, .concat_all = right_concat_all};
 					
 					query_graph.addPair(ral::cache::kpair(child->kernel_unit, "output_a", parent->kernel_unit, "input_a", left_cache_machine_config));
 					query_graph.addPair(ral::cache::kpair(child->kernel_unit, "output_b", parent->kernel_unit, "input_b", right_cache_machine_config));
@@ -453,7 +470,7 @@ struct tree_processor {
 					}
 					
 					cache_settings cache_machine_config = cache_settings{.type = CacheType::CONCATENATING, .num_partitions = 1, .context = context->clone(),
-						.concat_cache_num_bytes = concat_cache_num_bytes, .concat_all = false};
+						.concat_cache_num_bytes = concat_cache_num_bytes, .num_bytes_timeout = concatenating_cache_num_bytes_timeout, .concat_all = false};
 					query_graph.addPair(ral::cache::kpair(child->kernel_unit, parent->kernel_unit, cache_machine_config));
 				} else {
 					cache_settings cache_machine_config;